Show exportstack errors in split UI
Summary:
If split fails to load, previously it would not surface the error anywhere, the modal would just never appear. Instead, we should show the modal but show the error instead of the content.

Granted, I don't know any case where this error actually trips, but it's good to be able to surface it even so.

Reviewed By: quark-zju

Differential Revision: D57401887

fbshipit-source-id: 7a811005662c7f625685cadf6c80c101f97693a9