Show split/edit stack modals immediately in loading state, show waiting for ongoing commands
Summary:
After the previous diff, we now wait to open the split modal until your current operation finishes. Without updates, this is a bad UX because it would "silently" wait for a potentially long amount of time.

So it's better if we show the split modal with a loading spinner during this time. In fact, I think it's better to immediatley show the split / edit stack UIs while they load, even if they're just loading the regular stack state and not waiting for commands to finish.

So here while edit stack or split are loading, we immediatley show a modal with a loading spinner. We add the relevant classnames so the height won't change during and after loading (this makes the loading state less obtrusive).

Also, we add a dismiss button in the corner during the loading state. This way, if it's "stuck" loading, you're not stuck in that state. Or if you realize it's waiting for too many commands for a long time, you can close out.

Reviewed By: quark-zju

Differential Revision: D57395179

fbshipit-source-id: 06e7178edd790b8cd7ede7a71669e393f0874b28