Extract function to get all changed files for a hash
Summary:
Currently, we fetch the first 25 files using `sl log`. So, some APIs need a way to get more files than this.

In the near future, we will likely make this 25 file fetch go to 0 files fetched by default, and require all changes to be fetched through this function. This allows the `sl log` call to skip fetching files (which aren't needed most of the time), to improve fetch performance. This especially helps for transparent git.

Reviewed By: quark-zju

Differential Revision: D57458041

fbshipit-source-id: f4ff9cf41a69a46bcf69402561e88eafef00220a