Add FilesSample type with filesSample and totalFileCount, use when fetching files
Summary:
Use a dedicated type with a subset of all files, combined with the count of all files. This way, the UI can know the total number of files but only render or use a subset, and also know when it's using a subset or if it knows all the files.

This matches what already exists in `CommitInfo`, but we can use this when fetching just files without the rest of the CommitInfo.

Reviewed By: quark-zju

Differential Revision: D57458042

fbshipit-source-id: 8bde18f0ce6de7d550a901d59d831b2ed678a791