Add limit to commit's files being fetched
Summary:
Add a limit parameter to file fetching so we can reduce fetching in pathological cases, like commits with 20,000 files changed. Almost no ISL use case needs to have all of these files, so we can make do with the first 1000 and add a confirmation step anytime you want more than that.

This diff just adds a limit parameters. We'll still need to:
- rename the message to not e "fetchAll" since it's not "all" anymore
- change the data structure we store so that we keep the number of files fetched. This will let us use the cache as long as the number of files already fetched is >= the number being requested now. This will make it possible to request more files

Reviewed By: quark-zju

Differential Revision: D57458040

fbshipit-source-id: 249f3bf0c88d31d72de8a29fb51127ad98f7e0c4