status: filter results from eden using ignore matcher
Summary:
In D57171618 I took the ignore matcher out of the `matcher` object. The assumption was the underlying "filesystems" (physical, watchman, eden) already respect the ignore matcher. That is true for eden, except that eden does not know about global ignore files configured in the sapling config (e.g. "ui.ignore.my_ignore=~/my_ignore"). This caused ignored files to sneak by as untracked files. Fix by filtering "added" files from eden through our ignore matcher.

A better solution is to consolidate the global ignore configuration so sapling and eden are always in sync, but that will have to be a future project.

Reviewed By: quark-zju

Differential Revision: D57457079

fbshipit-source-id: 6317433a23de9036f77d1111c05f8653f70210dd