Add "BothChanged" Sync case
Summary:
Previously, in cases where both remote and local have changed, we'd classify the sync situation as "LocalIsNewer" for simplicity. But we can classify this more specifically as "BothChanged", which will allow us to show a new UX, which can acknowledge that there's a new verison to download, but that your local version has also changed during that time.

Detecting this is unfortunately not perfect without looking at ALL predecessors, which could be expensive to lookup. So we approximate this feature to the first order by only looking at the closest single predecessor via information we already have in `CommitInfo.closestPredecessor`.

This means we can only detect BothChanged accurately if you have exactly one local change. If you rebase AND amend, or rebase twice, etc, it would not be able to differentiate between LocalIsNewer and BothChanged, and so we fall back to LocalIsNewer. This means BothChanged is "best effort" only and not guaranteed. The other sync states are guaranteed.

{F1640351668}

This diff detects this case, but does not do anything about it. In the next diffs we can use this to make the UI descriptions and actions more accurate.

Reviewed By: jsaindon

Differential Revision: D57469924

fbshipit-source-id: 4fb5625ddb5da2b6d7bf10edb5eabbf33dfb3b15