Use BothChanged sync state to show special Download button
Summary:
Check for the `BothChanged` state, and use it to show the "Download new version" button, as if it was in `RemoteIsNewer` state. But add extra handling so we show a special tooltip and confirmation interstitial. We want to make the user aware that this new version will require some manual attention to bring into sync.

Ideally we'd have a way to compare the versions, so you can see why exactly they've diverged. Most of the time, you've probably just rebased or something and it's not significat changes. In those cases, a quick comparison would be enough to tell you which side you should use. But we don't have a good 3-way merge component yet.

You can technically use "Compare with phabricator" (which still works), but it's a little clunky because it shows your local copy as undoing the remote changes.

Reviewed By: jsaindon

Differential Revision: D57469925

fbshipit-source-id: f611af9f1863ffd347a1db979553c6d05bd2121d