Mononoke API: expose affected changesets limit
Summary:
Allow the caller of the API to configure the affected changesets limit for bookmark create and move operations.
This limit ensures that we don't load an unbounded number of changesets during bookmark operations.

This is currently a no-op change as we stick to the previous hard-coded limit of 200k commits.

This plumbing change will allow the next diff to intentionally expose this limit to gitimport, allowing us
to import very large git repos such as chromium/src which currently hit that limit.

Reviewed By: mitrandir77

Differential Revision: D57278832

fbshipit-source-id: 18945d5800afa132ed4efdd7a19f7dac107f6f92