Improve git_locator host selection logic for Mononoke GRit hosts
Summary:
The `git_locator` crate is used by Metagit (replicating follower) and in location service to redirect traffic to Mononoke GRit server instead of regular Metagit hosts. Currently, we always select the first host that SR returns us. Additionally, we just return the host IP and port and not the host name. This makes it difficult to debug in case of errors.

This diff updates the API of `git_locator` crate so we can make better choices in host selection (in follow up diffs) and also provide more context on the selected host

Reviewed By: mzr

Differential Revision: D57382486

fbshipit-source-id: c4f7e8711c28abdf16a2ce89b9f97b2a39dec1e8