Make fields get split by `,` more consistently
Summary:
Tokenized fields in commit messages are stored as comma-separated strings, but often parsed back into arrays for use in the UI. It seems this was a little inconsistent, and it was possible to get `"myToken,"` with a trailing comma as a field value. It doesn't seem there were any buggy consequences of this, only that you might see an extra comma in some places.

It's more robust if we are consistent about tokens not containing commas, and using commas only in the underlying string representation.

This diff fixes a few small errors, most notably when saving array token fields as `[value]` instead of `value.split(',')`.

This has the effect of making the field conflict preview from "Fill commit message" not having trailing commas in the display.

Reviewed By: bhamodi

Differential Revision: D57513189

fbshipit-source-id: 613c805ae1feccf6530036f3bf1d1d0424a0a27a