Fix commit info fields not always tokenizing the last token
Summary:
The way tokenization of fields in the commit info view was being handled, when you click to edit a field, it would be inconsistent about whether it made the last token into a token or left it editable in the editor. I think it's better to make it a token so you don't need to press enter again to add more tokens after that. You can always press delete and retype the token if you need to modify it.

If the underlying message happened to have a trailing comma, then it would do this already. So the goal here is just to make this more consistent.

This requires:
- not changing the underlying message because we may still get messages from outside ISL that don't seet the trailing comma
- when you start editing a field, insert a blank empty field at the end in the edited message
- make sure empty trailing fields don't count as meaningful edits to the message so it doesn't confirm when you click "cancel" if you don't make any changes.

Reviewed By: bhamodi

Differential Revision: D57509691

fbshipit-source-id: 07c900bc6fe64973dbff8bc91fa3f818bc3f3435