switch the main verification function from source/target to large/small
Summary:
The verify_working_copy_with_version function had basically two behaviour variants based on whether the function was provided with small->large or large->small which in the end were supposed to function in the same way.

This diff simplifies the logic by making the function first identify which repo is large and small and then just use those terms to proceed forward in consistent way.

Reviewed By: gustavoavena

Differential Revision: D57165973

fbshipit-source-id: 53e4fc7af77cdb9c3767dbc633c2d651f2475aec