gitimport: Allow branches of any size
Summary:
Specify the limit on how many changesets away from public changesets the bookmarks can be to be the number of commits in the repo.
This means that we are willing to pay the runtime cost of running hooks on all changesets for any size of a repo.
This seems preferable to skipping the check (which could be done by manipulating the config: `run_hooks_on_additional_changesets`, in case
some changesets would be problematic. We'd rather see the gitimport fail in that case and react accordingly.

This allows us to import repos where one branch is longer than 200k commits.

For instance, that is the case of `chromium/src`.

Reviewed By: mitrandir77

Differential Revision: D57278831

fbshipit-source-id: a75cf154bfe770bfb3ab82b61d844db213c6eab0