setup: use dummy version to make setuptools happy
Summary:
The version number we generate with `make oss` caused an error with the
setuptools. Since sapling is not a typical Python library - it is not intended
to be used in other Python projects (ex. put `sapling == 1.0` in some other
python project's `requirements.txt`), we can just use a dummy version to make
setuptools happy.

Reviewed By: zzl0

Differential Revision: D57573424

fbshipit-source-id: 8da284418b459ec84be832eafe41dbd40844975c