Update `sl url` to use cwd if no files are provided
Summary:
If you don't pass any files, `sl url` would just print nothing, which isn't exactly useful. Let's just have it use the cwd as the file. This way, you can use `sl url` and it prints a url for the current dir.

This is also useful for ISL, which can use this to get a url for the repo at a particular commit, without needing to be in the context of asking about a particular file, but the repo itsefl.

Reviewed By: quark-zju

Differential Revision: D57630150

fbshipit-source-id: a11d21d5ab1e8f4344ad4aee979221ee7346ccf9