Support dangling submodule pointers in small repo config
Summary:
## Problem
Some git repos have in their history submodule pointers to commits that don't exist.
This is a problem that comes with using git submodules.

Currently, if we're not able to find a commit in the submodule repo, we crash.

## Solution
Because this mostly happens in the repo's history, it's only a problem during initial import. Let's add a new field to the small repo sync config for submodule pointers that we know are dangling.

When such a pointer is found, instead of crashing, we'll create a changeset in the submodule repo containing a README file informing that this corresponds to a dangling submodule pointer.

This approach allows us to still detect unexpected dangling submodule pointers while correctly handling known ones.

## This diff
This diff starts using this `dangling_submodule_pointers` field to do this.

This diff replaces all calls to `bonsai_git_mapping` in the submodule expansion code with the new function `get_submodule_bonsai_changeset_id`. This will look for the git commit in bonsai_git_mapping, but if it doesn't find it, it will check the dangling pointers list and will create the dummy changeset if required.

Differential Revision: D57621151

fbshipit-source-id: 7098e0570ac3e4032e3578e37bf6b6ff2507aa42