Add menu item to copy file url
Summary:
Add a menu item on changed files to copy the url to the file. Like the previous diff, this depends on the `fbcodereview.code-browser-url` config.

If the config is available, then we can run `sl url` with the file path and the relevant comparison revset to get a URL we can send to the client to copy to the clipboard.

Reviewed By: quark-zju

Differential Revision: D57634161

fbshipit-source-id: 651b9cf56c53771eea8010622b22d30a5284789d