browse repo at hash
Summary:
Use `sl url` to allow opening a link on a public commit to the repo at that commit.

This is only enabled if `fbcodereview.code-browser-url` config is set for the repo.

I currently only show this option on public commits. Draft commits already have lots of context menu actions, and I think this is perhaps most useful for a public base.

Note we need D57630150 for this to work well, so that if your cwd is the repo root, it correctly still returns the url.

Reviewed By: quark-zju

Differential Revision: D57633285

fbshipit-source-id: 08d7e25af879895a5618b007226f58768c855e14