Add "Save all" button to save all unsaved files
Summary:
In the context menu next to the count of unsaved files, add a "Save all" choice, which just saves all the files.

The client sends a message to the server to do the save, and also sends a message afterwards about the success. This will be useful in the next diff, where we confiirm on commit/amend if you want to save unsaved files before committ/amend.

Reviewed By: quark-zju

Differential Revision: D57681441

fbshipit-source-id: 24a283405d62214c56a917514234abef973b65a9