Show unsaved files in vscode
Summary:
Start showing an indicator if you have some unsaved files in open vscode editors. This can help you remember to commit/amend those changes.

This feature could be implemented by other IDEs too, so it's managed by platform-specific messages.

In the next diff, we can add a menu to see which files are unsaved, open them, and even save them. We'll also check this state when amending/committing to ask if you want to save the files first

Reviewed By: quark-zju

Differential Revision: D57679994

fbshipit-source-id: f298d7033cba249a9d3ba76db97a2c34aa789519