Confirm before committing/amending with unsaved changes
Summary:
Before running commit/amend, if you have unsaved changes, ask if you want to save them first. This is a simple reminder with quick actions to save all or continue without saving. It also shows which files are unsaved.

This will only have an effect in vscode (currently) where we know which files are unsaved.

Reviewed By: quark-zju

Differential Revision: D57683159

fbshipit-source-id: f73c98057974381a17029f43907b75e9b3d0d712