eden: Add handle.exe output parser in Rust
Summary:
handle.exe is the tool we use in Windows to find which processes are holding on to resources.

For the Python part of Eden we have code that uses it, but we don't have an equivalent for Rust, and we need it to kill processes that might prevent redirect fixup to work.

This diff adds a parser for the output of handle.exe and its tests, but it doesn't yet use it.

Reviewed By: kmancini

Differential Revision: D56842465

fbshipit-source-id: e160b4ab1405729f14e06fa4dbd2d6402cc34495