eden: redirect fixup --force
Summary:
This diff introduces the option --force to fixup

When used, fixup will attempt to remove directories even if they have contents.

This is the first part of the work on this front.

Next diff will attempt to recover from failures in case files are locked by processes holding on to them.

Reviewed By: kmancini

Differential Revision: D55594044

fbshipit-source-id: 1b1a923831ab7f5c38898b8300e259556fb1e372