Deshim atomic_read_mostly_main_ptr in eden
Summary:
The following rules were deshimmed:
```
//folly/experimental:atomic_read_mostly_main_ptr -> //folly/concurrency/memory:atomic_read_mostly_main_ptr
```

The following headers were deshimmed:
```
folly/experimental/AtomicReadMostlyMainPtr.h -> folly/concurrency/memory/AtomicReadMostlyMainPtr.h
```

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

Reviewed By: kmancini

Differential Revision: D57681718

fbshipit-source-id: 2ce6927a52b5832097a97bde03202dcd9fb13520