use existing generated files logic to improve exlusion of gen'd files in sloc calculations
Summary:
continuing to iterate on the pro tip to split diffs.  this diff is improving the exclusion of gen'd files in the SLOC calculations by using the preexisting gen'd file checks that exist.  for now I'm just using the filesSample which is limited to 25 files.  future iterations may paginate through the files until we can determine the diff has +100 lines changed.

additionally, for now we are only doing this check on diffs with < 25 files (total) as there could be perf issues but also more than likely changes w/ huge # of files are likely to be codemods etc.

Reviewed By: quark-zju

Differential Revision: D57785433

fbshipit-source-id: 009bb973b32858384f33c97d1398291aaf6c8814