Fallback to small repo blobstore in InMemoryRepo reads
Summary:
## Problem
The logs show that a significant amount of CPU time during the initial-import is spent copying file content blobs between small repo, submodule repos and large repo.

## Context
There are 2 places we copy in the `expand.rs` module. This diff deals with the first one (seems to be most problematic):

where we copy all blobs from the expansion from small repo to large repo blobstore before finishing the expansion logic.

We do this **because the large repo needs to have the file contents from the new expansion files to derive fsnodes for expansion validation**.

## Solution
This copy can be avoided by modifying the `InMemoryRepo` to take fallback repos which the inner repo (i.e. large repo) will fallback to when reading from the blobstore. We then pass the small repo and all the submodule dependencies as the fallback repos.

This way, we don't need to keep copying blobs in advance.

Differential Revision: D57722367

fbshipit-source-id: f81e0f81b362785fbf1f46bb54f5cc58cb80791a