Backfill derivation of synced ancestors
Summary:
## Problem
Since D57665857, we derive all data types for the imported commits in batches. If the initial import is killed or crashes (e.g. db errors P1371278331), the latest commits will likely not have been derived. This means that the next run will likely fail because it will call `derive_batch` on the first batch, which expects to have all ancestors derived.

I've been mitigating this problem by manually backfilling all synced ancestors that weren't derived using newadmin, **but this is pretty inconvenient** and the level of automation of the process, bc I need to run the command and **wait for it to finish before I can run the initial-import command again**.

## Solution
- When getting the unsynced ancestors, also get the latest synced ancestor.
- Use `derived_utils` to derive all data types for that changeset. It will derive everything for all its ancestors that haven't been derived as well.

This might not be the most efficient way to do this, but (a) it will be done only before starting the import and (b) is definitely better than what I'm currently doing.

Differential Revision: D57697987

fbshipit-source-id: 521112016eb47737f908b86bd11c54166ed61687