Prevent delta cycles due to filemode change in Mononoke GRit server
Summary:
When deriving `GitDeltaManifest`, we perform `manifest_diff` on the current commit's `GitTree` with the parent commit's `GitTree`. This is supposed to give us everything that got added, removed or modified in the current commit.

Every change that we come across as a result of this manifest_diff is expressed as a delta (as long as it fits the criteria) with the underlying assumption that there is a change in content. However, the user can simply update the filemode of a file (e.g. by making it executable) which doesn't update the content but still shows up as a change in the response of `manifest_diff`. In such a case, we try to express the same version of a file (with hash H) as a delta of the previous version of this file (also with hash H). This obviously leads to delta cycle. This problem was discovered when performing incremental pulls of `oculus/aosp/vendor/oculus` repo.

This diff performs an additional check when deriving GitDeltaManifest to ensure that we don't include delta entries where the base and target are the same hash.

NOTE: This diff itself will not retroactively fix all the `GitDeltaManifest` derivations for past commits. This is strictly a forward fix which would especially be important for `GitDeltaManifestV2`

Differential Revision: D57858806

fbshipit-source-id: 2bde3d20647b7954d1009b5453f8d08276b91f60