Prevent self-referential deltas in Mononoke GRit server
Summary: In D57858806, we fixed the potential for creating deltas in `GitDeltaManifest` where the base object is the same as the target object (i.e. a self referential delta). However, this will fix all manifests that will get generated in the future and won't address the existing ones. For all such cases, we can modify the packfile stream generator in git protocol to skip over such self-referential deltas to prevent git erroring out due to delta cycles.

Reviewed By: gustavoavena

Differential Revision: D57872707

fbshipit-source-id: 5209b2b279fd2d0d0921aef19afb92439c20d14f