Define new struct for commit rewrite results
Summary:
This diff is a no-op change before D57764277, which improves the speed of syncing commits with submodule expansions.
This has been very helpful during initial import runs with large repos.

## What it does
Introduces a struct to store the result of `rewrite_commit`.

## Why?
Because in the next diffs `rewrite_commit` will start returning a with all content ids that need to be copied from submodule repos.

Differential Revision: D57907912

fbshipit-source-id: 2f9c1477af8053355f53a42a12afbfae3a0284a7