affected_changesets: Only load additional changesets once
Summary:
Move the loading of additional changesets to `check_restrictions`, if any check is needed.

This is more functional. It allows to remove the mutability from each individual check and makes the data flow one step clearer.

This makes the OOM fixing at the end of this stack more straightforward.

Reviewed By: gustavoavena

Differential Revision: D57906017

fbshipit-source-id: 7bb720fff430b547f3ab4d4c8816a287899b3db5