affected_changesets: Only load and check each changeset once
Summary:
In the parent diff (D57906014), we stopped loading all changesets at once in memory to solve the problem of OOMs.

One problem is that this may mean we have to load the same changesets many times since we're not keeping that state in memory.

Instead, store the hashes of already checked changesets and make sure to never reload them again.

We are still storing up to `O(repo)` data, but it's hashes instead of full changesets, so the memory cost of this optimization
is `O(repo)` instead of `O(repo)*changeset size`.

This shouldn't OOM in the context of gitimport where the largest repos contain a few million commits.

In other contexts, we don't expect branches to be very far away from public commits. We still have a hard-coded limit at 200_000 changesets too for these cases.

Reviewed By: gustavoavena

Differential Revision: D57906011

fbshipit-source-id: 1ebc0949225cf50795f4db758157a7b0a585dbd0