affected_changesets: Don't rely on pre-loading of changesets
Summary:
In a previous diff in this stack (D57906017), we ensured that we are just loading the changesets once per bookmark operation.
This means we don't need to try and save work by verifying whether the changesets were already loaded or not.

In this stack, we want to become lazy about loading the changesets, which means that this behaviour wouldn't work anyway.

Reviewed By: gustavoavena

Differential Revision: D57906013

fbshipit-source-id: 6bdf68455b1c75f8974e388f5aecf068a7282538