affected_changesets: Only load a bounded number of changesets into memory at once
Summary:
Before this change, all changesets seen by this bookmark that needed to be checked was loaded into a giant `HashSet` and processed at once.

In the case of `gitimport`, this would lead to loading `O(n commits in repo)` changesets into memory at once, leading to OOMs for large repos such as `chromium/src`, even with >220 GB of RAM.

After this change, we only load them by chunks of `10_000`, leading to a bounded amount of memory loaded at once.

Note that this is also a runtime pessimization when creating many bookmarks as we will reload all changesets each time for the checks.

Reviewed By: gustavoavena

Differential Revision: D57906014

fbshipit-source-id: d264e737d75a69ec1d2547a600429a64ebe68a9e