bisect: implement a more efficient bisect algorithm
Summary:
The current bisect algorithm does not play well with lazy changelog. It has a
O(bisect-range) compleixty to download commit hashes, and antoher O(bisect-range)
for testing a lazy skip revset.

The new algorithm avoids the O(bisect-range) complexity:
- Use fast native dag algorithms to avoid downloading the commit hashes as much
  as possible.
- If a lazy "skip" revset is used, look commits around the "best" and use early
  return. Its worst case is still O(biset-range), but if the skip set is small
  then hopefully a better practical UX.

This is based on a workaround script I worte in 2022 (P496941136), updated to
support the lazy "skip" revset.

Reviewed By: zzl0

Differential Revision: D57888302

fbshipit-source-id: 212dfcfad708065e24d2048886b7596c516b1825