io: add method to reset terminal
Summary:
Add a method to reset the terminal (and call it from the ctrl-c handler). This should help get the terminal back to a good state after when you ctrl-c out of Python readline operations (or other things that change terminal settings).

In particular, we make sure the cursor is visible and reset termios settings. This is a subset of what termwiz UnixTerminal does to reset the terminal (when dropped).

Reviewed By: quark-zju

Differential Revision: D57890781

fbshipit-source-id: 482d72ae8d471ae1807ea8a827308da8daf61740