Add `NotAvailable` SubmoduleDeps variant
Summary:
Introduce an enum variant to SubmoduleDeps to specify when any submodule dependencies were not able to be loaded.

This will be used in the next diff (D57913718) to stop failing open when submodule dependencies are not available during RepoContext creation.

Currently the new variant  will behave the same as `NotAvailable` in all callsites, but I prefer to err on the side of being more granular and using the type to differentiate each scenario.

Reviewed By: RajivTS

Differential Revision: D57914680

fbshipit-source-id: 4bd7f926ef8e76252d8d1e5b900eb678fe3c6fab