Extend integration test with repo that expands submodules
Summary:
Extend the existing integration test that sets up multiple repo syncs to a large repo to have a new forward synced repo which expands submodules.

NOTE: I reused components from the submodule integration tests, but their output is not really relevant in this test, so I dumped it to files to avoid making this (already long) test even longer.

Reviewed By: mitrandir77

Differential Revision: D57059292

fbshipit-source-id: 4b3f8c1377f23d84a64e8ff71b02483401c181ed