git_move_bookmark: Use MononokeScubaSampleBuilder
Summary:
This gives us the ability to specify the scuba dataset with `"file://some/path/to/a/file.json"` to log scuba samples to a file.

This is useful in integration tests.

NOTE: We don't want to pull-in many Mononoke dependencies in `git_move_bookmark` as it needs to be small and fast to be used in the context of a git hook. It's OK to do this because `MononokeScubaSampleBuilder` is a very low-dependency crate.

Reviewed By: RajivTS

Differential Revision: D58010114

fbshipit-source-id: 21d30d846963327d43d3b9c2dbe14aead71c709e