Enable `eden doctor` to warn when running on bad versions
Summary:
## Background
When using EdenFS on windows, it is easier to stay on a bad version due to the inability to automatically restart eden.

## Solution
Add bad version list config in doctor config along with reason and warn user as part of checks in doctor when running on one of these bad version.

Also recommend users to use `edenfsctl restart --graceful` to migrate to the newer version.

## Tracking
[Scuba](https://fburl.com/scuba/edenfs_events/vbqf76f6)

Reviewed By: kmancini

Differential Revision: D57696055

fbshipit-source-id: 5acf3dd293d9ab6c9727d1653c77fd98655a0576