add expected identity to MononokeThriftClient
Summary:
Here I'm going with a bit different approach that with `scsc`. Making this
value configurable in Moononoke configs properly would take lots of effort so
since this will be used only in integration tests and is not part of any binary
we're distributing to our users let's set it via environment variable.

Reviewed By: quark-zju

Differential Revision: D58021909

fbshipit-source-id: a92879b92869e6ef7243c80c50647dc4d2b0abcd