Fix live batched import watches
Summary:
liveBatched watches are designed to record the number of batches and the max duration of batch run. In the `prepareRequests` function, we only add a watch in `liveBatchedBlobWatches_`. Therefore we don't have any watch in `liveBatchedTreeWatches_` or `liveBatchedBlobMetaWatches_`

This diff uses `requestType` enum to determine which liveBatched watch list should get a new watch.

Reviewed By: kmancini

Differential Revision: D57705197

fbshipit-source-id: ef5a54b977f3b05aa547c5e5cbe04173661c0e84