add new ODS signals to OverlayStats
Summary:
Overlay layer has durations for all of its load and save dir functions. However, we need to add two more durations for File load and save.
Also, there are few counters setup in this layer. We need more counters in this layer. Because when a function is stuck in Overlay, the duration will not end and record data. Sometimes in debugging it is worth looking at the Overlay counters to find whether Overlay layer is working or not?

Reviewed By: kmancini

Differential Revision: D57238585

fbshipit-source-id: 9fc9cf2eb76ddcad3ce511660837c23205feb274