debug: alias debugruntest to ".t"
Summary:
This makes it easier to type - `sl .t` to run a test. The `.` does not conflict
with other end-user-facing commands.

Initially, I thought about aliasing `debug*` to `_*` (D56851668). Later, I
realized it does not have to be all `debug` commands - high frequent debug
commands are a small subset, and it does not have to be `_`.
`.t` matches the file extension more closely.

Reviewed By: sggutier

Differential Revision: D57927407

fbshipit-source-id: b732588039f0320b4e236bd7d80779a09893d193