testing: add a way to record states of a test
Summary:
Conceptually similar to D22007501 (for run-tests.py), track the state of test
run so they can be "replayed" later. I used to manually truncate a test, re-run
it to check its state. This command tracks all states for all "$" commands so
there is no need to manually truncate tests.

Note our tests now use `.sl` and `.git` in TESTTMP. So we can no longer just
use Git like D22007501 to track files. Additionally, this new implementation
also tracks the exported env vars and cwd, which D22007501 doesn't.

I thought about using EagerRepo instead of MetaLog, which can be useful to
allow using `sl` commands against the EagerRepo to query states (ex. what files
changed after this command in the test). But it might be tricky to deal with
the `.sl` and `.git` special names, and the EagerRepo's APIs are relatively
"raw" at present. So metalog for now. Maybe in the future we can switch to
EagerRepo.

Reviewed By: muirdm

Differential Revision: D57927411

fbshipit-source-id: 8eb4e4eb1679dc9d93aac219b00b07c60aebf07f