mononoke_types: add MPathElementPrefix
Summary:
Add a new type to represent *prefixes* of path elements.  This will be useful for traversing sharded manifests where sometimes we need to represent partial path elements.

This is another wrapper around `SmallVec<[u8; 24]>`.  The key difference is that some validation of `MPathElement` does not apply: prefixes can be empty and are also permitted to have values like `.` and `..`, as these are still valid as prefixes.

Reviewed By: liubov-dmitrieva

Differential Revision: D57965530

fbshipit-source-id: adeefad5276dc58357d2da340334527709e68e13