Open non generated files
Summary:
When clicking "Open all files" in the commit info view, it would open generated files too. This is almost always not what you want, IMO.

This diff changes the behavior to open only non-generated files by default.

If some files are generated, the button changes to "open non-generated files" for clarity.

If all files are generated, then, to avoid the button being a no-op, the button changes back to "open all files", and the tooltip clarifies that all files are generated. I also considered hiding or disabling the button in this case but we might as well give you some utility for it.

I also think if we want to allow opening generated files then we could have a second button "Open generated files" in the "Generated files" collapsable section. But I'm not sure this is actually useful.

Reviewed By: quark-zju

Differential Revision: D58033317

fbshipit-source-id: cf24b8591134462b2a833983766b758a0781503d