Autofocus fields in the commit info view when you click to edit them
Summary:
This was a very annoying bug. If you clicked to start editing a field, it would not get focus in all cases. You'd have to click again, or keyboard focus it.

The reason for this is that we attempt to make the topmost field be "autofocused", to prevent focusing the bottom field when we start editing many fields at once.

The solution in this diff is to auto-focus the topmost field of the set of fields that have just started being edited during this render.

For example, if no fields are edited, then you click one field, the topmost of the set is just the field that you clicked on, so it is focused.

If no fields are edited then you click "amend as..." to edit ALL fields, it should only autofocus the topmost field (the title)

If you edit some fields, then click on another commit and click back, it will reload the fields to be edited. The topmost of those fields that you've previously edited will be autofocused.

This is accomplished with `usePrevious` added in the last diff. This makes it easy to tell what fields have just started being edited.

Reviewed By: quark-zju

Differential Revision: D58116914

fbshipit-source-id: 49338bd22b7504efbee4eecfc7dd1702a432068b