Add usePrevious hook
Summary:
This hook keeps the last different value for any given input. This lets you store and detect changes to a value, so you can take action on it.

This would be an antipattern if used instead of useEffect cleanup return values. But in some cases, it's needed to diff and take actions.

The motivating use case for this is to auto-focus the topmost edited field in the commit info view. It should find which fields have started being edited just now, and choose which to render with "autofocus", so the single topmost field can get edited. This ends up working well with this simple abstraction.

Reviewed By: quark-zju

Differential Revision: D58116915

fbshipit-source-id: 00e8ccad8cb30c23727eac8dedddce110c6047a2