commit graph: Add ancestors_difference_segment_slices method
Summary:
This provides a convenient way to iterate over the commit graph in topological order by slices of segments.

It returns a stream of slices of commits, which allows us to only load one slice at a time and avoid excessive memory consumption.

This is useful, for instance in D58019789 to iterate over the commit graph by slices of commits that are well behaved for the purpose of batch derivation.

Reviewed By: mitrandir77

Differential Revision: D58135190

fbshipit-source-id: 6a322fcac0aecb4cf8544b05ef247b2d4fda2d28