scs: Relax invariants for calling repo_prepare_commits
Summary:
Before this, we could only call that endpoint if all parents of the batch which we wanted derived was already derived.
We also would fail if the batch wasn't topologically sorted.

With this change, we can tolerate a wider set of inputs: we will find what's missing and derive it in batches of 1000 commits.

This will allows us to be more lenient around the live git sync.

We will still need to exercise some caution to avoid causing `O(repo)` work during the live sync, hanging a push; but that's not new.

Reviewed By: mitrandir77

Differential Revision: D58019789

fbshipit-source-id: a188f126de6746b7509fa321d5fcc89fb999ecd9